fix: address review comment 4590903660 - merge methods
