+++ tests/integration/runtime/test_pr_monitor_runner_parts/test_pr_monitor_runner_part_003.py
@@ -36,7 +36,7 @@ from awf.runtime.pr_monitor_runner import (
 )
 from awf.runtime.pr_monitor_runner.helpers import _parse_verdict
 from tests.postgres import postgres_test_engine
-from tests.unit.runtime._monitor_runner_fixtures import DefaultMergeMethodGitHubClient
+from tests.shared.monitor_runner import DefaultMergeMethodGitHubClient
 
 
 @dataclass
