+++ tests/shared/__init__.py
@@ -1,0 +1,1 @@
+"""Shared test helpers used across AWF test layers."""
